Don't require OkBuffer callers to cast.

+++ okhttp-tests/src/test/java/com/squareup/okhttp/internal/spdy/HpackDraft05Test.java
@@ -800,7 +800,7 @@ public class HpackDraft05Test {
 
   private void assertBytes(int... bytes) {
     ByteString expected = intArrayToByteArray(bytes);
-    ByteString actual = bytesOut.readByteString((int) bytesOut.size());
+    ByteString actual = bytesOut.readByteString(bytesOut.size());
     assertEquals(expected, actual);
   }
